Polish up Cabrillo output. (#4)
In detail: The output of attributes is order-independent according to
the spec.  But sanity demands that the same software outputs the same
stuff in a consistent order.  This was not the case previously, as
plain Python dict had been used and had been iterated over, resulting
in ever-changing output produced from run to run.  But consistency
has been achieved with this change.

The Cabrillo spec demands: "All QSO lines must appear in chronological
order."  Now the Cabrillo timestamp granularity is only one minute,
and QSO rates above 60 per hour are not at all uncommon.  So (strictly
speaking), chronological order cannot be achieved by sorting QSOs via
timestamp.  It is somewhat unclear whether anybody would actually care
to receive a file with QSO lines not in chronological order, but only
sorted by Cabrillo timestamps.

We take a conservative stance and never re-order QSO data.  We assume
that incoming data already is ordered chronologically.

This required a move of functionality: The distinction between valid
QSO data and X-QSO data used to be done by using different lists.
Now, one single list is used, and the QSO itself knows whether it
is valid (QSO:) or not (X-QSO:).

We also check, on input, whether incoming data is indeed ordered
chronologically, as the Cabrillo spec demands and as far as can be
checked, given the coarse timestamps.  On disorder found, an exception
is raised.  If you do not need ordering, this check can be switched
off.  If you do switch it off, Cabrillo output generation is switched
off as well, as there is a high risk the output might not be valid
Cabrillo.